# openapi.yaml
...
openapi: 3.1.0
info:
  title: VoiceAI engine + platform + telephony
  version: 0.10.217
  description: 'Merged spec: engine (agent CRUD, /health, ws /chat/v1/{agent_id})
    + platform routers (/calls/simulate, /executions/stats, batches, auth, graphs,
    workflows) + telephony (/talko/*, /twilio/*, /plivo/*). Generated from FastAPI
    openapi() (A11 hygiene 2026-09-12).'
paths:
...
  /plivo/call:
    post:
      tags:
      - Plivo Telephony
      summary: Initiate Outbound Call (Plivo)
      description: Initiates an outbound call using Plivo to the specified recipient
        phone number and connects it to the specified agent.
      operationId: make_call_call_post
      parameters:
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CallDetails'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallInitiated'
        '401':
          description: Missing or invalid X-API-Key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Invalid request body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: Plivo rejected the call.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: ngrok or Plivo credentials unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /plivo/plivo_connect:
    post:
      tags:
      - Plivo Telephony
      summary: Plivo Answer URL Callback
      description: Callback endpoint for Plivo to provide XML instructions for streaming
        audio to the VoiceAI WebSocket server.
      operationId: plivo_connect_plivo_connect_post
      parameters:
      - name: voiceai_host
        in: query
        required: true
        schema:
          type: string
          description: The public URL of the VoiceAI websocket host
          title: Voiceai Host
        description: The public URL of the VoiceAI websocket host
      - name: agent_id
        in: query
        required: true
        schema:
          type: string
          description: The ID of the agent to connect
          title: Agent Id
        description: The ID of the agent to connect
      responses:
        '200':
          description: XML instructions returned successfully.
          content:
            application/json:
              schema: {}
        '403':
          description: Plivo signature invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '400':
          description: VOICE_STREAM_SECRET not configured.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ErrorBody:
      properties:
        code:
          type: string
          title: Code
          description: Stable error code; see voiceai.errors.ErrorCode.
        message:
          type: string
          title: Message
          description: Caller-safe description of what went wrong.
        error_id:
          type: string
          title: Error Id
          description: Correlation id that also appears in the server log.
        retryable:
          type: boolean
          title: Retryable
          description: Whether the same request may succeed if retried.
          default: false
        component:
          anyOf:
          - type: string
          - type: 'null'
          title: Component
          description: Subsystem that failed (llm, synthesizer, api, ...).
        provider:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider
          description: Upstream provider, when the failure was theirs.
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
          description: Model or deployment involved, when known.
        details:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Details
          description: Structured context such as a config path or validation issues.
      type: object
      required:
      - code
      - message
      - error_id
      title: ErrorBody
      description: Structured error as it appears on the wire (``error`` key).
    ErrorEnvelope:
      properties:
        ok:
          type: boolean
          title: Ok
          description: Always false for error bodies.
          default: false
        detail:
          type: string
          title: Detail
          description: Same as error.message; kept for FastAPI-style clients.
        error:
          $ref: '#/components/schemas/ErrorBody'
      type: object
      required:
      - detail
      - error
      title: ErrorEnvelope
      description: Body of every non-2xx HTTP response produced by the engine.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    CallDetails:
      properties:
        agent_id:
          type: string
          minLength: 1
          title: Agent Id
          description: The ID of the agent to handle the call.
        recipient_phone_number:
          type: string
          minLength: 3
          title: Recipient Phone Number
          description: The phone number to call in E.164 format (e.g., +1234567890).
      type: object
      required:
      - agent_id
      - recipient_phone_number
      title: CallDetails
    CallInitiated:
      properties:
        status:
          type: string
          title: Status
          default: initiated
        request_uuid:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Uuid
          description: Plivo request UUID when the carrier accepted the request.
        agent_id:
          type: string
          title: Agent Id
      type: object
      required:
      - agent_id
      title: CallInitiated